--- a/torchao/_models/sam2/automatic_mask_generator.py
+++ b/torchao/_models/sam2/automatic_mask_generator.py
@@ -210,17 +210,16 @@ class SAM2AutomaticMaskGenerator:
             elif self.output_mode == "binary_mask":
                 mask_data["segmentations"] = [rle_to_mask(rle) for rle in mask_data["rles"]]
             else:
-                # import pdb; pdb.set_trace()
-                rles = []
-                # TODO: Using .cpu() directly plus unbind seems to cause weakref error.
-                rles_nt_cpu = torch.nested.nested_tensor_from_jagged(mask_data["rles_nt"].values().cpu(), mask_data["rles_nt"].offsets().cpu())
-                for (mask_i, d) in zip(rles_nt_cpu.unbind(), mask_data["rles_sizes"]):
-                    h, w = d["size"]
-                    if mask_i.size() == 0:
-                        rles.append({"size": [h, w], "counts": []})
-                    else:
-                        rles.append({"size": [h, w], "counts": [0] + (mask_i.tolist()[:-1])})
-                mask_data["rles"] = rles
+                # rles = []
+                # # TODO: Using .cpu() directly plus unbind seems to cause weakref error.
+                # rles_nt_cpu = torch.nested.nested_tensor_from_jagged(mask_data["rles_nt"].values().cpu(), mask_data["rles_nt"].offsets().cpu())
+                # for (mask_i, d) in zip(rles_nt_cpu.unbind(), mask_data["rles_sizes"]):
+                #     h, w = d["size"]
+                #     if mask_i.size() == 0:
+                #         rles.append({"size": [h, w], "counts": []})
+                #     else:
+                #         rles.append({"size": [h, w], "counts": [0] + (mask_i.tolist()[:-1])})
+                # mask_data["rles"] = rles
                 mask_data["segmentations"] = mask_data["rles"]
 
         with torch.autograd.profiler.record_function("_encode_masks: Write mask records"):
@@ -352,8 +351,8 @@ class SAM2AutomaticMaskGenerator:
         with torch.autograd.profiler.record_function("uncrop_points"):
             data["points"] = uncrop_points(data["points"], crop_box)
         with torch.autograd.profiler.record_function("crop_boxes"):
-            # data["crop_boxes"] = torch.tensor([crop_box for _ in range(len(data["rles"]))])
-            data["crop_boxes"] = torch.tensor([crop_box for _ in range(len(data["rles_nt"]))])
+            data["crop_boxes"] = torch.tensor([crop_box for _ in range(len(data["rles"]))])
+            # data["crop_boxes"] = torch.tensor([crop_box for _ in range(len(data["rles_nt"]))])
 
         return data
 
@@ -447,8 +446,8 @@ class SAM2AutomaticMaskGenerator:
                     # TODO: Capture all these masks in a single NT for mask_to_rle_pytorch_2
                     # or at a minimum create a mask_to_rle_pytorch_2_list and use loops
                     # to cause a single DtoH sync
-                    # data["rles"] = mask_to_rle_pytorch_2(data["masks"])
-                    data["rles_nt"] = mask_to_rle_pytorch_2_nt(data["masks"])
+                    data["rles"] = mask_to_rle_pytorch_2(data["masks"])
+                    # data["rles_nt"] = mask_to_rle_pytorch_2_nt(data["masks"])
                     del data["masks"]
 
                     batch_data = data
@@ -591,10 +590,10 @@ class SAM2AutomaticMaskGenerator:
         with torch.autograd.profiler.record_function("uncrop_masks"):
             # Compress to RLE
             data["masks"] = uncrop_masks(data["masks"], crop_box, orig_h, orig_w)
-            # data["rles"] = mask_to_rle_pytorch_2(data["masks"])
-            data["rles_nt"] = mask_to_rle_pytorch_2_nt(data["masks"])
-            b, h, w = data["masks"].size()
-            data["rles_sizes"] = [{"size": [h, w]}] * b
+            data["rles"] = mask_to_rle_pytorch_2(data["masks"])
+            # data["rles_nt"] = mask_to_rle_pytorch_2_nt(data["masks"])
+            # b, h, w = data["masks"].size()
+            # data["rles_sizes"] = [{"size": [h, w]}] * b
             del data["masks"]
 
         return data