--- a/torchao/_models/sam2/automatic_mask_generator.py
+++ b/torchao/_models/sam2/automatic_mask_generator.py
@@ -534,6 +534,7 @@ class SAM2AutomaticMaskGenerator:
                 # Filter by predicted IoU
                 if self.pred_iou_thresh > 0.0:
                     keep_mask = data["iou_preds"] > self.pred_iou_thresh
+                    # TODO: Might need this for correctness due to calculate_stability_score IoU?
                     # data.filter(keep_mask)
 
             with torch.autograd.profiler.record_function("calculate_stability_score"):