RLEData, some reorder of filtering, more compile-able regions. Causes slight numerical differences across all test images, but passes unit tests bit perfectly. Improves perf from ~190ms to ~130ms in the furious case with an increase in number of mismatched mask count. Improves perf from ~390ms to ~320ms without changing the mismatched mask count.